This commit refactors all floats into doubles.  It also moves get_color_contribution to Primitive.
As a consequence of moving get_color_contribution to Primitive, get_center was also moved to Primitive.
This was done to eliminate the need for Primitive to have knowledge of it's subclass Sphere, if that
would even be possible.  The consequence of this is that all primitives now supported the concept of
having a center.  This may prove necessary later anyway (without a center, where does the object
get rendered at?). 




git-svn-id: https://svn.code.sf.net/p/mbrt/code/trunk@22 a272ff57-b438-0410-9c92-d377374fc898

+++ color.h
@@ -7,13 +7,13 @@ typedef unsigned char uchar;
 //{{{
 class Color {
     public:
-        float red;
-        float green;
-        float blue;
+        double red;
+        double green;
+        double blue;
 
         Color() : red(0.0f), green(0.0f), blue(0.0f) {}
-        Color(float _red, float _green, float _blue) : red(_red), green(_green), blue(_blue) {
-            //std::cout << "float ctor" << std::endl;
+        Color(double _red, double _green, double _blue) : red(_red), green(_green), blue(_blue) {
+            //std::cout << "double ctor" << std::endl;
             set_red(red);
             set_green(green);
             set_blue(blue);
@@ -48,13 +48,13 @@ class Color {
         }
         Color operator* (const Color &other) { return Color(red * other.red, green * other.green, blue * other.blue); }
 
-        void operator*=(float n) {
+        void operator*=(double n) {
             //std::cout << "operator*=(Color)" << std::endl;
             set_red(red * n);
             set_green(green * n);
             set_blue(blue * n);
         }
-        Color operator* (float n) { return Color(red * n, green * n, blue * n); }
+        Color operator* (double n) { return Color(red * n, green * n, blue * n); }
 
         void operator+=(const Color &other) {
             //std::cout << "operator+=(Color)" << std::endl;
@@ -66,9 +66,9 @@ class Color {
         Color operator+ (const Color &other) { return Color( red + other.red, green + other.green, blue + other.blue); }
         Color operator-(const Color &other) { return Color(red - other.red, green - other.green, blue - other.blue); }
 
-    void set_red (float new_red) { red = (new_red < 1.0f) ? new_red : 1.0f; /*std::cout << "red = " << red << std::endl;*/ }
-    void set_green (float new_green) { green = (new_green < 1.0f) ? new_green : 1.0f; /*std::cout << "green = " << green << std::endl;*/ }
-    void set_blue (float new_blue) { blue = (new_blue < 1.0f) ? new_blue : 1.0f; /*std::cout << "blue = " << blue << std::endl;*/ }
+    void set_red (double new_red) { red = (new_red < 1.0f) ? new_red : 1.0f; /*std::cout << "red = " << red << std::endl;*/ }
+    void set_green (double new_green) { green = (new_green < 1.0f) ? new_green : 1.0f; /*std::cout << "green = " << green << std::endl;*/ }
+    void set_blue (double new_blue) { blue = (new_blue < 1.0f) ? new_blue : 1.0f; /*std::cout << "blue = " << blue << std::endl;*/ }
 
         uchar get_rgb_red () const { return (uchar)(red * 255); }
         uchar get_rgb_green() const { return (uchar)(green * 255); }
