+++ point3d.h
@@ -10,17 +10,17 @@ typedef Point3D Vector;
 //{{{
 class Point3D {
     public:
-        float x;
-        float y;
-        float z;
+        double x;
+        double y;
+        double z;
 
         Point3D() : x(0.0f), y(0.0f), z(0.0f) {}
 
         template <class _T>
-        Point3D(_T _x, _T _y) : x((float)_x), y((float)_y), z(0.0f) {}
+        Point3D(_T _x, _T _y) : x((double)_x), y((double)_y), z(0.0f) {}
 
         template <class _T>
-        Point3D(_T _x, _T _y, _T _z) : x((float)_x), y((float)_y), z((float)_z) {}
+        Point3D(_T _x, _T _y, _T _z) : x((double)_x), y((double)_y), z((double)_z) {}
 
         Point3D(const Point3D &other) : x(other.x), y(other.y), z(other.z) {}
 
@@ -30,7 +30,7 @@ class Point3D {
         /// @f[ d=\sqrt{(\Delta x)^2+(\Delta y)^2+(\Delta z)^2}=\sqrt{(x_1-x_2)^2+(y_1-y_2)^2+(z_1-z_2)^2} @f]
         ///
         /// @param other The other point to which we are determining the distance.
-        inline float distance_to(const Point3D &other);
+        inline double distance_to(const Point3D &other);
 
 
         std::string to_string() {
@@ -41,14 +41,14 @@ class Point3D {
 
         template <class _T>
         void set(_T x, _T y, _T z) {
-            this->x = (float)x;
-            this->y = (float)y;
-            this->z = (float)z;
+            this->x = (double)x;
+            this->y = (double)y;
+            this->z = (double)z;
         }
 
         /// Calculate the normalized vector from the origin to the destination.
         void normalize() {
-            float norm = sqrtf(x * x + y * y + z * z);
+            double norm = sqrtf(x * x + y * y + z * z);
             x /= norm;
             y /= norm;
             z /= norm;
@@ -59,10 +59,10 @@ class Point3D {
 
 // This is inline, it must be defined in the header.
 //{{{
-float Point3D::distance_to(const Point3D &other) {
-    float dx = x - other.x;
-    float dy = y - other.y;
-    float dz = z - other.z;
+double Point3D::distance_to(const Point3D &other) {
+    double dx = x - other.x;
+    double dy = y - other.y;
+    double dz = z - other.z;
     return sqrtf(dx*dx + dy*dy + dz*dz);
 }
 //}}}
@@ -87,7 +87,7 @@ inline Vector operator+(const Vector &A, const Vector &B) {
 }
 //}}}
 //{{{
-inline Vector operator*(const Vector &A, float factor) {
+inline Vector operator*(const Vector &A, double factor) {
     return Vector(A.x * factor, A.y * factor, A.z * factor);
 }
 //}}}
@@ -105,14 +105,14 @@ inline bool operator!=(const Vector &A, const Vector &B) {
 
 /// The dot product of two unit vectors provides the cosine of the angle between the two vectors.
 //{{{
-inline float dot_product(const Vector &A, const Vector &B) {
+inline double dot_product(const Vector &A, const Vector &B) {
     return (A.x * B.x) + (A.y * B.y) + (A.z * B.z);
 }
 //}}}
 //{{{
 template <class _T>
 inline Vector operator*(Vector &v, const _T n) {
-    return Vector(v.x * (float)n, v.y * (float)n, v.z * (float)n);
+    return Vector(v.x * (double)n, v.y * (double)n, v.z * (double)n);
 }
 //}}}
